+++ R/auk-date.r
@@ -8,10 +8,17 @@
 #' @param x `auk_ebd` or `auk_sampling` object; reference to file created by 
 #'   [auk_ebd()] or [auk_sampling()].
 #' @param date character or date; date range to filter by, provided either as a
-#'   character vector in the format `"2015-12-31"` or a vector of Date objects.
+#'   character vector in the format `"2015-12-31"` or a vector of Date objects. 
+#'   To filter on a range of dates, regardless of year, use `"*"` in place of 
+#'   the year.
 #' 
-#' @details This function can also work with on an `auk_sampling` object if the 
-#'   user only wishes to filter the sampling event data.
+#' @details To select observations from a range of dates, regardless of year, 
+#' the  wildcard `"*"` can be used in place of the year. For example, using 
+#' `date = c("*-05-01", "*-06-30")` will return observations from May and June 
+#' of *any year*.
+#' 
+#' This function can also work with on an `auk_sampling` object if the user only 
+#' wishes to filter the sampling event data.
 #'
 #' @return An `auk_ebd` object.
 #' @export
@@ -23,6 +30,12 @@
 #' # alternatively, without pipes
 #' ebd <- auk_ebd(system.file("extdata/ebd-sample.txt", package = "auk"))
 #' auk_date(ebd, date = c("2010-01-01", "2010-12-31"))
+#' 
+#' # the * wildcard can be used in place of year to select dates from all years
+#' system.file("extdata/ebd-sample.txt", package = "auk") %>%
+#'   auk_ebd() %>%
+#'   # may-june records from all years
+#'   auk_date(date = c("*-05-01", "*-06-30"))
 auk_date <- function(x, date)  {
   UseMethod("auk_date")
 }
@@ -35,11 +48,20 @@ auk_date.auk_ebd <- function(x, date) {
     is.character(date) || assertthat::is.date(date),
     date[1] <= date[2]
   )
+  
+  # check for wildcard in year
+  has_wildcard <- stringr::str_detect(date, "^\\*-[0-9]{1,2}-[0-9]{1,2}")
+  if (all(has_wildcard)) {
+    # temporarily replace wildcard with 2016
+    date <- stringr::str_replace(date, "^\\*", "2016")
+  } else if (!all(!has_wildcard)) {
+    stop("Cannot mix wildcard dates with non-wildcard dates.")
+  }
 
   # convert to date object, then format as ISO standard date format
   date <- as.Date(date) %>%
     format("%Y-%m-%d")
-
+  
   assertthat::assert_that(
     all(!is.na(date)),
     date[1] <= date[2],
@@ -48,7 +70,14 @@ auk_date.auk_ebd <- function(x, date) {
   )
 
   # define filter
-  x$filters$date <- date
+  if (all(has_wildcard)) {
+    x$filters$date <- stringr::str_replace(date, "^2016", "*")
+    attr(x$filters$date, "wildcard") <- TRUE
+  } else {
+    x$filters$date <- date
+    attr(x$filters$date, "wildcard") <- FALSE
+  }
+  
   return(x)
 }